refactor: move backup utilities to owa.core.utils
- Move BackupContext and DummyConsole from owa.cli.mcap to owa.core.utils.backup
- Update all import statements across codebase to use new location
- Move tests from owa-cli to owa-core for better organization
- Remove exports from owa.core.utils.__init__.py to keep it clean

BREAKING CHANGE: BackupContext import path changed from owa.cli.mcap.backup_utils to owa.core.utils.backup
