+++ projects/owa-cli/owa/cli/mcap/migrate/rollback.py
@@ -14,7 +14,6 @@ from typing import List, Optional
 import typer
 from rich.console import Console
 
-from owa.cli.mcap.backup_utils import BackupContext
 from owa.cli.mcap.migrate.file_utils import (
     add_rollback_row,
     create_file_info_table,
@@ -22,6 +21,7 @@ from owa.cli.mcap.migrate.file_utils import (
     format_datetime,
     format_file_size,
 )
+from owa.core.utils.backup import BackupContext
 
 
 @dataclass
